fix order and employee services

+++ src/main/java/com/my/bookstore/security/JwtUtils.java
@@ -20,7 +20,7 @@ public class JwtUtils {
     private String jwtSecret;
 
     @Value("${jwt.expiration}")
-    private int jwtExpiration;
+    private long jwtExpiration;
 
     private SecretKey secretKey;
